Fix parser error with certain demo files
Some demo files would fail due to user data being present when the
parser did not expect it to be. The issue was that when the
'string_table_update' event was fired it was passing the current
user_data field of the string table entry being updated to the
callback function.

During each string table update a bit is read from the buffer to
determine whether or not to read new user data. If that flag is not
set then there is no new user data. The correct thing to do is
either pass the new user data (when the flag is True) or to pass None
(when the flag is False) to any callback functions for the
'string_table_update' event.