Add basic serializers for entities
The way entity classes are determined has also been changed. Previously
it relied on the class ID of the incoming entity. However, this
doesn't seem to be stable as different demos sometimes have different
class IDs. Instead the entity class will be determined based on the
presence of data tables. E.g., if an entity baseline has a 'DT_Team'
data table then it is assumed to be a Team entity.